# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Quasar Compiler API",
    "description": "API criada para compilar arquivos da QuasarLanguage",
    "version": "1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/files/upload": {
      "post": {
        "tags": [
          "api/files"
        ],
        "description": "Realiza o upload de arquivos",
        "operationId": "uploadFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Download realizado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Erro ao realizar o download",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/list": {
      "get": {
        "tags": [
          "api/files"
        ],
        "description": "Realiza a listagem de arquivos enviados e gerados",
        "operationId": "listFiles",
        "responses": {
          "200": {
            "description": "Listagem realizada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao realizar a listagem",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/files/download/{fileName}": {
      "get": {
        "tags": [
          "api/files"
        ],
        "description": "Realiza o download de arquivos",
        "operationId": "downloadFile",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download realizado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "500": {
            "description": "Erro ao realizar o download",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {

  }
}